ENH: Sends the type of exceptions back along with the serialized object.
To differentiate between `ValueError('v')` and `raise ValueError('v'),
qdb sent back a bool in the `exc` field that meant: was an exception
thrown. This made clients rely on a consistent exception serializer to
derive the type.

Now the tracer will send back the type of the exception raised in the
`exc` field of `print`, `watchlist`, and `cond error` events, or `None`
if there was no exception raised.

When a condition for a breakpoint raises an uncaught exception, the
message has been changed to include the type in the `exc` field, and the
serialized exception is passed in the new `output` field. This
standardizes all three cases where exceptions are passed back, making
`exc` always mean the same thing.