BUG: Config file heirarchy was not respected because there was no means of saying, "defer to the next level".
Implements a Maybe monad where QdbConfigs operate on Maybes and convert
to strict values when all composition is done.

To denote that I want to defer this input to the next highest level of
configuration, you can mark the key with a Nothing()
object (singleton). This is the default for all keys not explicitly set.

Values highest on the stack of configurations will have their values
used, unless it is Nothing, in which case it will defer down the stack
until a Just is found.
This order is:

- kwargs
- user-files (in order passed)
- local config file
- profile config file
- default values (These are always Just)